+++ thrifty-gradle-plugin/src/main/kotlin/jp/co/gahojin/thrifty/gradle/KotlinThriftOptions.kt
@@ -64,6 +64,9 @@ open class KotlinThriftOptions : ThriftOptions(), Serializable {
     @get:Input
     var isGenerateServer: Boolean = false
 
+    @get:Input
+    var jvmOverloads: Boolean = false
+
     fun setServiceClientStyle(clientStyleName: String) {
         val stylesByName = TreeMap<String, ClientStyle>(String.CASE_INSENSITIVE_ORDER)
         for (style in ClientStyle.entries) {